# specs/001-coaching-velo-garmin-strava/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: Coaching vélo connecté — API
  version: 0.1.0
  description: >
    Contrat d'API pour l'import de séances (Garmin Connect / Strava / Nolio),
    l'analyse de charge d'entraînement, et les recommandations de récupération
    et nutrition. Couvre US1, US2, US3 du spec.
paths:
...
  /api/connexions/{plateforme}/autoriser:
    post:
      summary: Démarrer le flux d'autorisation OAuth pour une plateforme (FR-001)
      parameters:
        - name: plateforme
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Plateforme"
      responses:
        "200":
          description: URL de redirection vers le flux d'autorisation de la plateforme
          content:
            application/json:
              schema:
                type: object
                required: [url_autorisation]
                properties:
                  url_autorisation:
                    type: string
                    format: uri
  /api/connexions/{plateforme}/callback:
    post:
      summary: Callback OAuth — finalise la connexion après autorisation (FR-001)
      parameters:
        - name: plateforme
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Plateforme"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "201":
          description: Connexion établie, import initial déclenché (SC-001)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnexionPlateforme"
  /api/connexions/{plateforme}:
    delete:
      summary: Déconnecter une plateforme (révoque les tokens localement)
      parameters:
        - name: plateforme
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Plateforme"
      responses:
        "204":
          description: Connexion révoquée
...
components:
  schemas:
    Plateforme:
      type: string
      enum: [garmin_connect, strava, nolio]
    StatutConnexion:
      type: string
      enum: [actif, expire, revoque]
...
    ConnexionPlateforme:
      type: object
      required: [id, plateforme, statut]
      properties:
        id:
          type: string
          format: uuid
        plateforme:
          $ref: "#/components/schemas/Plateforme"
        statut:
          $ref: "#/components/schemas/StatutConnexion"
        date_derniere_synchronisation:
          type: string
          format: date-time
          nullable: true
        date_connexion:
          type: string
          format: date-time
...